Revert "Auction authorization"
