+++ website/api_urls/auction.py
@@ -5,6 +5,7 @@ from website.views import (
     MemberAuctionList,
     CategoryAuctionList,
     SubcategoryAuctionList,
+    AuctionDeleteByUser,
 )
 
 urlpatterns = [
@@ -19,19 +20,23 @@ urlpatterns = [
         name="auction-detail",
     ),
     path(
-        "user/<str:username>",
+        "User/<str:username>",
         MemberAuctionList.as_view({"get": "list"}),
         name="member-auction-list",
     ),
     path(
-        "category/<slug:name>",
+        "Category/<slug:name>",
         CategoryAuctionList.as_view(),
         name="category-auction-list",
     ),
     path(
-        "subcategory/<slug:subcategory_name>",
+        "Subcategory/<slug:subcategory_name>",
         SubcategoryAuctionList.as_view(),
         name="subcategory-auction-list",
     ),
-
+    path(
+        "delete/User/<str:username>",
+        AuctionDeleteByUser.as_view(),
+        name="auction-delete-by-user",
+    ),
 ]
