+++ website/tests/test_model.py
@@ -1,0 +1,6 @@
+from django.test import TestCase
+
+
+class TestModel(TestCase):
+    def test_model(self):
+        self.assertEqual(1, 1)
