+++ website/views/auction.py
@@ -1,48 +1,17 @@
-from rest_framework import generics, status
-from rest_framework.response import Response
+from rest_framework import generics
 from rest_framework.viewsets import ModelViewSet
 from website.serializer import (AuctionListSerializer,
                                 AuctionDetailSerializer,
                                 AuctionCreateSerializer,)
 
 from website.models import Auction
-from rest_framework.permissions import BasePermission
 
 # Auction views
-
-
-class AuctionPermission(BasePermission):
-    def has_permission(self, request, view):
-        if view.action == "list":
-            return True
-        elif view.action == "create":
-            return request.user.is_authenticated
-
-
-class AuctionEditPermission(BasePermission):
-    def has_permission(self, request, view):
-        # staff and superuser can do anything
-        if request.user.is_staff or request.user.is_superuser:
-            return True
-        if request.method == "GET":
-            return True
-        return False  # you should not be able to edit any auctions
-
-    def has_object_permission(self, request, view, obj):
-        # staff and superuser can do anything
-        if request.user.is_staff or request.user.is_superuser:
-            return True
-        if request.method == "GET":
-            return True
-        return False  # you should not be able to edit any auctions
-        # if you want to be able to edit your auctions, use this:
-        # return obj.auctionOwner == request.user
 
 
 class AuctionList(ModelViewSet):
     queryset = Auction.objects.all()
     serializer_class = AuctionListSerializer
-    permission_classes = (AuctionPermission, )
 
     def get_serializer_class(self):
         if self.action == "create":
@@ -51,8 +20,6 @@ class AuctionList(ModelViewSet):
 
 
 class AuctionDetail(generics.RetrieveUpdateDestroyAPIView):
-    permission_classes = (AuctionEditPermission, )
-
     queryset = Auction.objects.all()
     serializer_class = AuctionDetailSerializer
     lookup_field = "auctionID"
@@ -74,17 +41,10 @@ class SubcategoryAuctionList(generics.ListAPIView):
         return Auction.objects.filter(subcategory__subcategory_name=subcat)
 
 
-class MemberAuctionList(ModelViewSet):
+class AuctionDeleteByUser (generics.DestroyAPIView):
     queryset = Auction.objects.all()
-    serializer_class = AuctionListSerializer
-    permission_classes = (AuctionEditPermission, )
+    serializer_class = AuctionDetailSerializer
 
     def get_queryset(self):
         username = self.kwargs["username"]
         return Auction.objects.filter(auctionOwner__username=username)
-
-    def delete(self, request, *args, **kwargs):
-        username = self.kwargs["username"]
-        resp_data = Auction.objects.filter(
-            auctionOwner__username=username).delete()
-        return Response(resp_data, status=status.HTTP_204_NO_CONTENT)
